Using Liskov Substitution principle - we should not have subclass that does not extend all the behaviours of the parent class. Making rules as part of the Board class, defining ruleSet class to make the code more readable and extensible
