+++ src/main/java/game/GameConfig.java
@@ -1,4 +1,6 @@
 package game;
 
 public class GameConfig {
+    public Integer timePerMove;
+    boolean timed = false;
 }
