+++ src/main/java/game/Player.java
@@ -1,8 +1,9 @@
 package game;
 
 public class Player {
+    private int timeUsedInMillis;
     private String playerSymbol;
-
+    private User Id;
     public Player(String playerSymbol) {
         this.playerSymbol = playerSymbol;
     }
@@ -18,5 +19,13 @@ public class Player {
     public Player flip() {
         String flipPlayerSymbol = playerSymbol.equals( "X" )? "O":"X";
         return new Player( flipPlayerSymbol );
+    }
+
+    public void setTimeTaken( int timeUsedInMillis ){
+        this.timeUsedInMillis += timeUsedInMillis;
+    }
+
+    public int getTimeUsedInMillis() {
+        return timeUsedInMillis;
     }
 }
