+++ src/main/java/game/User.java
@@ -1,0 +1,5 @@
+package game;
+
+public class User {
+    private String ID;
+}
