+++ src/main/java/gamerules/Rule.java
@@ -1,0 +1,14 @@
+package gamerules;
+
+import game.Board;
+import game.GameState;
+
+import java.util.function.Function;
+
+public class Rule<T extends Board>{
+        Function< T, GameState > condition;
+        public Rule( Function< T, GameState > condition )
+        {
+            this.condition = condition;
+        }
+    }
